pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "cryocaustik", email = "alex@alazy.dev"},
]
dependencies = [
    "discord-py>=2.3.2",
    "python-dotenv>=1.0.0",
    "aiohttp==3.9.0b0",
    "requests>=2.31.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.11.0",
]
